Error in shotgun script
There is an error that I can't seem to fix, : error CS1061: 'Camera' does not contain a definition for 'tranform' and no accessible extension method 'tranform' accepting a first argument of type 'Camera' could be found (are you missing a using directive or an assembly reference?). If someone could fix the ray casting that would be great.